requirements.txt
asgiref==3.8.1
confusable-homoglyphs==3.3.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.2
django-registration==3.4
sqlparse==0.5.0
